requirements.txt
pyyaml
pandas
numpy
matplotlib
seaborn
tqdm
pillow
pathlib
tensorflow
scikit-image
scikit-learn
torch
torchvision
opencv-contrib-python